requirements.txt
torch>=1.10.0
torchvision>=0.11.1
transformers>=4.15.0
Pillow>=8.4.0
requests>=2.26.0
matplotlib>=3.5.0
gradio>=3.19.0
numpy>=1.21.0